fix: actions on click, rename folder
