+++ src/config/contexts.tsx
@@ -1,22 +1,24 @@
-import { OverlayProvider } from "./../common/context/OverlayContext";
+import { action } from "@storybook/addon-actions";
+import { OverlayProvider } from "../common/contexts/OverlayContext";
 
+// https://www.npmjs.com/package/@storybook/addon-contexts
 export const contexts = [
   {
     icon: "box",
-    title: "Overlay",
+    title: "Overlays",
     components: [OverlayProvider],
     params: [
       {
-        name: "OverlayContext",
+        name: "Overlay",
         props: {
           value: {
             overlayId: "",
             setOverlayId: () => {
-              console.log("set id");
+              action("Set overlay ID")();
             },
             isOverlayVisible: false,
             setOverlayVisible: () => {
-              console.log("set visible");
+              action("Close overlay")(); // https://github.com/storybookjs/storybook/issues/1153#issuecomment-442165039
             },
           },
         },
